+++ src/Commands/search/screenshot.js
@@ -1,11 +1,12 @@
-const { isUrlSafe, isUrl } = require('../../Structures/Util');
-const { MessageAttachment } = require('discord.js');
+const { isUrl } = require('../../Structures/Util');
 const { Command } = require('discord.js-commando');
+const request = require('node-superfetch');
 const WebShot = require('websshot');
 const window = new WebShot(
 	{ removeTags: [], removeAttributes: [], args: [], height: 800, width: 1280 },
 	{ args: ['--no-sandbox'], headless: true },
 );
+const url = require('url');
 
 module.exports = class ScreenshotCommand extends Command {
 	constructor(client) {
@@ -28,19 +29,37 @@ module.exports = class ScreenshotCommand extends Command {
 				},
 			],
 		});
+
+		this.nsfwList = null;
 	}
 
 	async run(msg, { link }) {
-		const safe = await isUrlSafe(link);
+		try {
+			if (!this.nsfwList) await this.fetchNSFWList();
 
-		if (!safe && !msg.channel.nsfw && msg.guild)
-			return msg.say(
-				'The URL you provided is an unsafe website. Use this on a NSFW channel instead.',
-			);
+			const parsed = url.parse(link);
 
-		const img = await window.screenshot(link);
-		const image = new MessageAttachment(img);
+			if (this.nsfwList.some((nsfwURL) => parsed.host === nsfwURL) && !msg.channel.nsfw)
+				return msg.reply('This site is NSFW. Please try again in a NSFW channel.');
 
-		return msg.say(image);
+			const attachment = await window.screenshot(link);
+
+			return msg.say({ files: [{ attachment, name: 'screenshot.png' }] });
+		} catch {
+			return msg.reply("Couldn't find any results. Invalid URL?");
+		}
+	}
+
+	async fetchNSFWList(force = false) {
+		if (!force && this.nsfwList) return this.nsfwList;
+
+		const { text } = await request.get('https://github.com/blocklistproject/Lists/raw/master/porn.txt');
+
+		this.nsfwList = text
+			.split('\n')
+			.filter((site) => site && !site.startsWith('#'))
+			.map((site) => site.replace(/^(0.0.0.0)/, ''));
+
+		return this.nsfwList;
 	}
 };